Add exitwaiter pkg and make TestReciever cleanup blocking
this is a long time coming. The TestReceiver shuts itself down (and by
extension linked processes) in a t.Cleanup() function, but we weren't
waiting for the Exit signal to be delivered. This is important in tests
of name registration, for example. By waiting, we ensure that we cleanup
all well-behaving processes before the next test starts.