fix some issues with test receiver
There was a race condition with the `exiting` boolean that needed to be
wrapped in a mutex.

Some of the logic around validating `expect.Times` and `expect.AtMost` was incorrect. Importantly,
the `Times` expectation will only be verified at the end of the test.

A few failing tests that were using the deprecated `erl` testreceiver
are fixed now.