fix ordering issue with inbox.Inbox
I observed messages in a client app
being received by genservers out of order (they were sent to `self`
serially so really shouldn't have been possible). Removing the signal
channel from `process` and replacing with Inbox worked in most cases but
the erl.TestExt_WorkerFanOut and other low level tests were timing out.
Next move was to not use the `Receive` method on the `Inbox` and instead
make a goroutine in the process that populated the channel that `Runnable`
used. I'm observering intermittant failures of TestWhereIs_NameNotFound, but
I've suspected an issue with name registration so this is progress.

Oh the failing tests in the client app resolved as well.